+++ src/components/Filter.jsx
@@ -14,7 +14,7 @@ import {
   UCircle,
   JCircle,
   KCircle,
-} from "./shapes/circles";
+} from "./shapes/circles.jsx";
 
 // all the components we need for the visualizer to work will live on this compnent
 // base on the value of the state, it will render the value of the key of the obj
