+++ src/components/shapes/diamonds.jsx
